Rename page to path
as it is very generic, that the response may not necessarily be a
"page".
This function is kept because we still need a generic one here at
C world (for folks not using Qt, for example).
Besides, the last time I tried, I remember Qt Network failed to
build for Emscripten.
So, this is what the function is for.. we need the API to stillm
be uniform, AND, works on Emscripten too.
It is very generic that it should later be out of the Interchange
specific client code.